+++ client/src/components/AppErrorBoundary/ErrorCatcher.tsx
@@ -1,5 +1,4 @@
 import {useEffect} from 'react';
-import {useNavigate} from 'react-router-dom';
 
 import toast from '@hooks/useToast/toast';
 
@@ -7,10 +6,8 @@ import {useAppErrorStore} from '@store/appErrorStore';
 
 import {captureError} from '@utils/captureError';
 import isRequestError from '@utils/isRequestError';
-import SessionStorage from '@utils/SessionStorage';
 
 import {SERVER_ERROR_MESSAGES} from '@constants/errorMessage';
-import SESSION_STORAGE_KEYS from '@constants/sessionStorageKeys';
 
 const isPredictableError = (error: Error) => {
   if (isRequestError(error)) if (error.errorCode === 'INTERNAL_SERVER_ERROR') return false;
@@ -20,7 +17,6 @@ const isPredictableError = (error: Error) => {
 
 const ErrorCatcher = ({children}: React.PropsWithChildren) => {
   const {appError: error} = useAppErrorStore();
-  const navigate = useNavigate();
 
   useEffect(() => {
     if (!error) return;
@@ -34,25 +30,7 @@ const ErrorCatcher = ({children}: React.PropsWithChildren) => {
       position: 'bottom',
       bottom: '8rem',
     });
-
-    // 로그인 처리
-    if (
-      error.errorCode === 'TOKEN_NOT_FOUND' &&
-      !window.location.pathname.includes('/guest/login') &&
-      !window.location.pathname.includes('/member/login')
-    ) {
-      const createdByGuest = SessionStorage.get<boolean>(SESSION_STORAGE_KEYS.createdByGuest);
-      SessionStorage.set<string>(SESSION_STORAGE_KEYS.previousUrlForLogin, window.location.pathname);
-
-      const currentPath = window.location.pathname;
-
-      if (createdByGuest) {
-        navigate(`${currentPath}/guest/login`);
-      } else {
-        navigate(`${currentPath}/member/login`);
-      }
-    }
-  }, [error, navigate]);
+  }, [error]);
 
   return children;
 };
